+++ apps/web/pages/submissions/upload.tsx
@@ -1,5 +1,4 @@
 import Layout from '@components/Layout';
-import { checkURL } from '@server/utils/apiHelper';
 import { ReactElement, useEffect, useState } from 'react';
 import { AlertStatus } from '@chakra-ui/alert';
 import {
@@ -65,6 +64,16 @@ export default function Upload() {
       await utils.gameplay.invalidate();
     },
   });
+  const checkURL = (url: string): boolean => {
+    const p =
+      // eslint-disable-next-line max-len
+      /^(?:https?:\/\/)?(?:m\.|www\.)?(?:youtu\.be\/|youtube\.com\/(?:embed\/|v\/|watch\?v=|watch\?.+&v=))((\w|-){11})(?:\S+)?$/;
+
+    if (url.match(p)) {
+      return true;
+    }
+    return false;
+  };
 
   const { data: isDisabled } = trpc.site.getPageData.useQuery({
     name: 'upload',
